Refactors drug_service.rb and nocolor_drug_result.rb helper methods into field_collector.rb module for reusability and refactoring
